feat: reformulate pkgerrors to stderror
Promote usage of the standard library error handling APIs over
x/xerrors.
