+++ go.mod
@@ -3,6 +3,6 @@ module github.com/einride/goreview
 go 1.15
 
 require (
-	golang.org/x/tools v0.0.0-20200225230052-807dcd883420
+	golang.org/x/tools v0.0.0-20200831203904-5a2aa26beb65
 	gotest.tools/v3 v3.0.2
 )
