--- internal/passes/pkgerrors/testdata/src/a/b.go
@@ -1,9 +1,0 @@
-package a
-
-import (
-	"golang.org/x/xerrors" // want `use "github.com/pkg/errors" instead of "golang.org/x/xerrors"`
-)
-
-func IllegalImports() {
-	_ = xerrors.New("hey")
-}
