+++ internal/passes/stderror/analyzer_test.go
@@ -1,4 +1,4 @@
-package pkgerrors
+package stderror
 
 import (
 	"testing"
