+++ internal/passes/stderror/testdata/src/a/a.go
