+++ internal/passes/stderror/testdata/src/golang.org/x/xerrors/mock.go
